chore: pricing v4 (#14336)
* chore: pricing v4

* fix: test

* chore: remove unused code

* chore: update

Former-commit-id: ba27c8ef9ca8261868202852c50c93db49c2350b
Former-commit-id: e502541fe4c851888215831263e230ec8fc49568

+++ backend/enterprise/api/license.go
@@ -71,7 +71,7 @@ type LicenseService interface {
 	// GetEffectivePlan gets the effective plan.
 	GetEffectivePlan() api.PlanType
 	// GetPlanLimitValue gets the limit value for the plan.
-	GetPlanLimitValue(ctx context.Context, name PlanLimit) int64
+	GetPlanLimitValue(ctx context.Context, name PlanLimit) int
 	// GetInstanceLicenseCount returns the instance count limit for current subscription.
 	GetInstanceLicenseCount(ctx context.Context) int
 	// RefreshCache will invalidate and refresh the subscription cache.
